Fix call to Outlook, because time has passed
New company (Treyarch), different Outlook, probably new Windows version,
so everything is broken, because Windows, and time has passed. This
works from MSYS to pop open Outlook from Vim, with buffer, or selected
lines, addressed to the appropriate party (from the dictionary in my
vimlocal local.vim file, specific to this machine), which small, but
helpful messagse in the subject line and body.